requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
httpx>=0.24.0
pillow>=11.0.0
line-bot-sdk==3.9.0
openai>=1.30.0
typing-extensions>=4.0.0
